SZ Better database, DAO, JPA handling.
NOTE: Connection pool is now created using C3P0 allowing multiple connextions
