+++ note-manager-api/src/main/java/com/steelzack/titletextadder/api/dao/TitleDao.java
@@ -1,12 +1,14 @@
 package com.steelzack.titletextadder.api.dao;
 
 import com.steelzack.titletextadder.api.model.Title;
+import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.EntityTransaction;
 import javax.persistence.Persistence;
 
+@Repository
 public class TitleDao implements TitleDaoInterface<Title> {
     private EntityTransaction currentTransaction;
 
@@ -17,20 +19,13 @@ public class TitleDao implements TitleDaoInterface<Title> {
         entityManagerFactory = Persistence.createEntityManagerFactory("note.manager.app.jpa");
     }
 
-    public void closeCurrentSessionwithTransaction() {
-        currentTransaction.commit();
-    }
-
-
-    private EntityTransaction getCurrentTransaction() {
-        return currentTransaction;
-    }
-
     public void persist(Title entity) {
         entityManager = entityManagerFactory.createEntityManager();
         currentTransaction = entityManager.getTransaction();
         currentTransaction.begin();
         entityManager.persist(entity);
+        currentTransaction.commit();
+
     }
 
     public void update(Title entity) {
@@ -38,6 +33,8 @@ public class TitleDao implements TitleDaoInterface<Title> {
         currentTransaction = entityManager.getTransaction();
         currentTransaction.begin();
         entityManager.refresh(entity);
+        currentTransaction.commit();
+
     }
 
     public void delete(Title entity) {
@@ -45,6 +42,7 @@ public class TitleDao implements TitleDaoInterface<Title> {
         currentTransaction = entityManager.getTransaction();
         currentTransaction.begin();
         entityManager.remove(entity);
+        currentTransaction.commit();
     }
 
 }
