+++ note-manager-api/src/main/resources/META-INF/persistence.xml
@@ -19,6 +19,16 @@
             <property name="hibernate.hbm2ddl.auto" value="create"/>
             <property name="show_sql" value="false"/>
             <property name="hibernate.current_session_context_class" value="thread"/>
+
+            <property name="hibernate.connection.provider_class"
+                      value="org.hibernate.connection.C3P0ConnectionProvider" />
+
+            <property name="hibernate.c3p0.max_size" value="100" />
+            <property name="hibernate.c3p0.min_size" value="0" />
+            <property name="hibernate.c3p0.acquire_increment" value="1" />
+            <property name="hibernate.c3p0.idle_test_period" value="300" />
+            <property name="hibernate.c3p0.max_statements" value="0" />
+            <property name="hibernate.c3p0.timeout" value="100" />
         </properties>
     </persistence-unit>
 </persistence>
