+++ note-manager-api/src/main/webapp/WEB-INF/mvc-servlet.xml
@@ -9,7 +9,7 @@
         http://www.springframework.org/schema/mvc
         http://www.springframework.org/schema/mvc/spring-mvc-4.0.xsd">
 
-	<context:component-scan base-package="com.steelzack.titletextadder.api.controller" />
+	<context:component-scan base-package="com.steelzack.titletextadder.api.controller, com.steelzack.titletextadder.api.service, com.steelzack.titletextadder.api.dao, com.steelzack.titletextadder.api.model" />
 
 	<mvc:annotation-driven />
 </beans>
